requirements.txt
numpy==1.23.5
tensorflow==2.11.0
tensorflow-hub==0.12.0
scikit-learn==1.2.0
Pillow==9.3.0
tqdm==4.64.1
Flask==2.2.3
werkzeug==2.2.3
gunicorn==20.1.0
python-dotenv==1.0.0
flask-cors==3.0.10
